fix: restructure followingList component
