+++ src/App.jsx
@@ -2,6 +2,8 @@ import 'src/style/reset.scss';
 import 'src/style/base.scss';
 import { MainPage, RegisterPage, SignInPage, AdminPage } from 'src/pages';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import { UserProfile } from './components/UserProfile/UserProfile';
+import { FollowingList } from './components/FollowSection/FollowingList/FollowingList';
 
 // import { compileString } from "sass";
 
@@ -15,6 +17,28 @@ function App() {
 					<Route path='main' element={<MainPage />} />
 					<Route path='admin' element={<AdminPage />} />
 					<Route path='*' element={<RegisterPage />} />
+					<Route
+						path='user/self'
+						element={
+							<MainPage>
+								<UserProfile
+									name={`John Doe`}
+									account={`John Doe`}
+									intro={`Nulla Lorem mollit cupidatat irure. Laborum magna nulla duis ullamco cillum dolor`}
+									followerCounts={`${59}位`}
+									followingCounts={`${34}個`}
+								/>
+							</MainPage>
+						}
+					/>
+					<Route
+						path='following'
+						element={
+							<MainPage>
+								<FollowingList name={`John Doe`} />
+							</MainPage>
+						}
+					/>
 				</Routes>
 			</BrowserRouter>
 		</div>
