+++ src/components/Sidebar/Sidebar.jsx
@@ -5,28 +5,29 @@ import { ReactComponent as User } from 'src/assets/icons/user-outline.svg';
 import { ReactComponent as Cog } from 'src/assets/icons/cog-outline.svg';
 import { ReactComponent as Logout } from 'src/assets/icons/logout.svg';
 import { ButtonL } from 'src/components/buttons';
+import { Link } from 'react-router-dom';
 
-export const Sidebar = () => {
+export const Sidebar = ({ onSectionClick }) => {
 	return (
 		<div className={style.sidebarContainer}>
 			<div className={style.sidebarWrapper}>
-				<a href='' className={style.sidebarLogo}>
+				<Link className={style.sidebarLogo} to='/main'>
 					<Logo />
-				</a>
+				</Link>
 				<div className={style.sidebarItemWrapper}>
-					<a href='' className={style.sidebarItem}>
+					<Link className={style.sidebarItem} to='/main'>
 						<div className={style.iconContainer}>
 							<Home className={style.sidebarIcon} />
 						</div>
 						<h5 className={style.sidebarTitle}>首頁</h5>
-					</a>
-					<a href='' className={style.sidebarItem}>
+					</Link>
+					<Link className={style.sidebarItem} to='/user/self'>
 						<div className={style.iconContainer}>
 							<User className={style.sidebarUserIcon} />
 						</div>
 						<h5 className={style.sidebarTitle}>個人資料</h5>
-					</a>
-					<a href='' className={style.sidebarItem}>
+					</Link>
+					<a href='' className={style.sidebarItem} onClick={onSectionClick}>
 						<div className={style.iconContainer}>
 							<Cog className={style.sidebarIcon} />
 						</div>
@@ -35,10 +36,10 @@ export const Sidebar = () => {
 				</div>
 				<ButtonL text='推文' className={style.button} />
 			</div>
-			<a href='' className={style.logout}>
+			<Link href='' className={style.logout} to='/signin'>
 				<Logout />
 				<h5 className={style.logoutTitle}>登出</h5>
-			</a>
+			</Link>
 		</div>
 	);
 };
